Made utils folder for consts and functions
