+++ utils/consts.py
@@ -1,0 +1,5 @@
+MAX_RANDOM_ID_INT = int(2e9)
+MIN_RANDOM_ID_INT = int(-2e9)
+START_WRITE_POSITION_X = 30
+START_WRITE_POSITION_Y = 50
+BLACK_COLOR = (0, 0, 0)
